Separated data initialization from recommend program.
