+++ recommend.py
@@ -1,6 +1,5 @@
 import chromadb
 import pickle
-import csv
 from sklearn.feature_extraction.text import TfidfVectorizer
 from pathlib import Path
 
@@ -20,7 +19,7 @@ def get_chroma_db() -> chromadb.PersistentClient:
     Raises:
         None
     """
-    return chromadb.PersistentClient(path='chroma_db')
+    return chromadb.PersistentClient(path='chroma')
 
 def get_top_recs(
     user_input: str, 
@@ -68,90 +67,7 @@ def print_recs(recs: dict) -> None:
     for movie in recs['metadatas'][0]:
         print(movie['title'])
 
-def read_movie_data() -> tuple[list, list, list, list]:
-    """
-    Read movie data from a CSV file.
-
-    Args:
-        None
-
-    Returns:
-        tuple: A tuple containing the movie data.
-    
-    Raises:
-        None
-    """
-    ids = []
-    documents = []
-    titles = []
-    metadatas = []
-
-    with open(CSV_FILE, 'r', encoding='utf-8') as file:
-        reader = csv.DictReader(file)
-        for row in reader:
-            row_dict = dict(row)
-            
-            ids.append(row_dict['id'])
-            titles.append(row_dict['Title'])
-            metadatas.append({"title": row_dict['Title']})
-            
-            document = f"{row_dict['Title']} {row_dict['Genre1']} {row_dict['Genre2']} {row_dict['Genre3']} {row_dict['Plot']}"
-            documents.append(document)
-    
-    return ids, documents, titles, metadatas
-
-def create_embeddings(documents: list) -> tuple[list, TfidfVectorizer]:
-    """
-    Create embeddings for the documents.
-
-    Args:
-        documents (list): The list of documents.
-    
-    Returns:
-        list: The list of embeddings.
-    
-    Raises:
-        None
-    """
-    vectorizer = TfidfVectorizer()
-    vectorizer.fit(documents)
-    embeddings = vectorizer.transform(documents).toarray().tolist()
-
-    return embeddings, vectorizer
-
-def init_db(
-    db: chromadb.PersistentClient, 
-    collection: chromadb.Collection
-    ):
-    """
-    Initialize the database with movie data.
-
-    Args:
-        db (chromadb.PersistentClient): The ChromaDB client.
-        collection (chromadb.Collection): The ChromaDB collection.
-    
-    Returns:
-        TfidfVectorizer: The vectorizer used to transform the input.
-    
-    Raises:
-        None
-    """
-    ids, documents, titles, metadatas = read_movie_data()
-    embeddings, vectorizer = create_embeddings(documents)
-    
-    collection.add(
-        ids=ids,
-        documents=documents,
-        embeddings=embeddings,
-        metadatas=metadatas
-    )
-    
-    with open(VECTORIZER_FILE, 'wb') as f:
-        pickle.dump(vectorizer, f)
-    
-    return vectorizer
-
-def load_vectorizer() -> TfidfVectorizer:
+def get_vectorizer() -> TfidfVectorizer:
     """
     Load the saved vectorizer from disk.
 
@@ -199,21 +115,18 @@ def get_input() -> str:
 def main():
     """Main function to run the program."""
     try:
-        db = get_chroma_db()
-        
-        # Check if database and vectorizer exist
-        db_exists = Path('chroma_db').exists()
+        db_exists = Path('chroma').exists()
         vectorizer_exists = Path(VECTORIZER_FILE).exists()
 
-        if not db_exists or not vectorizer_exists:
-            collection = db.create_collection('movies')
-            vectorizer = init_db(db, collection)
+        if not db_exists and not vectorizer_exists:
+            print("Data is not initialized. Run 'python init.py' to initialize the movie data.")
         else:
-            vectorizer = load_vectorizer()
+            db = get_chroma_db()
+            vectorizer = get_vectorizer()
+            user_input = get_input()
 
-        user_input = get_input()
-        recs = get_top_recs(user_input, vectorizer)
-        print_recs(recs)
+            recs = get_top_recs(user_input, vectorizer)
+            print_recs(recs)
 
     except Exception as e:
         print(f"Error: {e}")
